Simple test of full roundtripping working
This implements/ports enough of the fetch API in terms of WASI Preview2 to get a simple script working that handles an incoming `FetchEvent`, does an outgoing request to another host, and sends a response using `FetchEvent#respondWith`.

A key missing part is support for bodies, which will come next.

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>