Initial commit to the js.wasm project
This commit contains an import of code from [Fastly's js-compute-runtime](https://github.com/fastly/js-compute-runtime/) project. The code is unmodified, except for files being organized in a manner that'll make diffs between this and the following commit easier to read.

Note that this commit as-is doesn't build, as it's an incomplete import which only contains those parts that are reused by the js.wasm project.