Add basic support for JS modules
Specifically, the top-level script can now be executed as a module instead of a classic JS script. Module source code can refer to other modules using `import`, as expected.

The support is "basic" because right now it only supports importing JS modules with paths that are absolute or relative to the top-level module. Support for built-in modules and more complex external module resolution mechanisms is planned as future work.