Port WPT harness from Fastly's JS Compute runtime (#19)
The harness works, and is integrated with CMake in that there's now a build target `wpt-runtime` that results in a working runtime for running WPT tests, and a CTest definition for running the test suite.

What doesn't work all too well is the code in the runtime itself: as expected, running the test suite immediately uncovered a bunch of crashes and other errors. Those will be addressed in other commits.