Support configuring enabled builtins using CMake options
This modifies the `add_builtin` CMake function to derive a CMake option from the name of the builtin, and only to only install the builtin if that option is set.

The README.md is also modified to document how to get a list of all available builtins, and whether they're enabled by default.

Note that this mechanism will automatically apply to CMake projects using StarlingMonkey, too.