Allow specifying a dependency in add_builtins()
By ensuring that dependencies are installed in ${CMAKE_BINARY_DIR},
rather than ${CMAKE_CURRENT_BINARY_DIR}, we can find the installed
include directories and libraries from projects using StarlingMonkey
as a subproject.

This also ensures that a builtin that depends on another target will
trigger the build of that target appropriately.

In addition, this replaces the LINK_LIBS add_builtins() parameter,
becoming a more general approach to declaring a dependency for a
built-in.