Make starling.wasm a usable component, and document it as such
This changes the current `starling.wasm` build target to instead be called `starling-raw.wasm` to reflect that it's not yet a usable thing. Additionally, the current `runtime-eval` target is renamed to `starling` and produces a new `starling.wasm` file, which this time is a component that can be used to run dynamically loaded code.

I also updated the documentation accordingly.

And finally, I fixed a bug in `CMakeLists.txt` introduced in #151 that broke the `runtime-eval` build target. That's the entire reason I fell down this rabbit hole 😬

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>