Check the response status when fetching the file.
If the file name defined in tests.json cannot be found, the fetch
function doesn’t throw an error. Instead, it writes the error into the
response. Without checking the response status, we proceed to evaluate
the script source, which in this case is:

`Error: {"error": {"code": 404, "message": "404"}}`

At this point, eval throws an Invalid Syntax error without indicating
that the file wasn't fetched successfully.