Fix bugs in handling multiple incoming bodies piped through a `TransFormStream` to an outgoing body (#171)
This was pretty gnarly to work through, but I think came out reasonably clean. The two key pieces are:
- delaying the first read from an incoming stream that's been piped to a `TransformStream` until the latter is actually read from, and hence doesn't have backpressure applied anymore. This part guarantees that when the read would happen, we have the full pipeline in place and can hand things off to the host API instead of handling anything in the JS `fetch` API implementation.
- properly integrating these two pieces, and only closing the incoming body's `ReadableStream` once the host API reports that it's been dealt with to completion.

Additionally, the following two changes are folded in:

* Cleanup: improve some functions dealing with `Request` `Response` object handles

It was pretty hard to follow when a caller would expect to have a valid handle, or when a `nullptr` would be acceptable.

* Handle writing chunks to outgoing bodies fully asynchronously

Before, any chunk would be written synchronously, with write operations to the outgoing body in a loop. That both meant that large chunks would block all script execution until they're done, and that things wouldn't necessarily always work.

To wit, the test added here doesn't pass without this change. It streams a sentence in a very convoluted way, sending nested requests for each word, which then respond by sending the word back, letter for letter. Without this patch, that part breaks after the second letter in a word has been sent back.

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>