Fix loading of WPT META scripts (#209)
WPT tests can load helper scripts using special comments at the top. These scripts need to be loaded as though using a `<script>` tag, which doesn't create a new lexical scope, and hence causes `let` and `const` bindings to be visible in the global scope.

Previously, to simulate this behavior using `eval`, the WPT harness replaced `let` and `const` with `var`. While this worked surprisingly well, it doesn't *always* work.

This commit instead introduces an `evalScript` function available to the harness, which evaluates in a way more closely matching that of script tags.

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>