Add platform support for JS content debugging (#218)
This commit introduces platform support for debugging JavaScript. "Platform support" here means that the PR doesn't contain an implementation of an actual debugger. Instead, it adds infrastructure to StarlingMonkey to initiate a debugging session over a TCP socket connection, and to load a JS implementation of an actual debugger via that connection.

The debugger itself works, but isn't quite ready for prime-time, so I'll open it as a draft PR, building on this one.

The debugger requires sort of a double opt-in: StarlingMonkey has to be started with the -d/--enable-script-debugging option set, and a port to connect to needs to be provided via the DEBUGGER_PORT env var. That ensures that StarlingMonkey will not try to read env vars at runtime without a prior opt-in, and hence will guarantee that it'll continue working in environments that stub out wasi::environment.