Change Rust integration to emit a single static library (#212)
Instead of compiling multiple crates to static libs and trying to link all of them into the final binary, we now synthesize a crate that bundles all individual crates into a single staticlib. We maintain a single `Cargo.lock` file for this synthesized crate, which means that we can have a situation in which not all dependencies are tracked. That's "okay" in that we can live with the lockfile being best-effort.

This also replaces linking SpiderMonkey's Rust static library with including the required crates into our own library, achieving an end result of having a single static Rust library to link, and enables thin-LTO for all libraries: this is in general useful, but even more relevant with more Rust code.

Rolled into this is an update to newer version of Corrosion: the release of Rustup 1.28.0 broke the version of Corrosion we were previously using, so an update is required regardless of whatever else we do with our Rust integration. Since the newer version of Corrosion didn't work out-of-the-box with the previous Rust setup, I decided to fold it in here.

And finally, the PR includes a bugfix in TextDecoder. The fixed issue is preexisting, but is only getting caught with the Rust changes. Either something changed in Rust's stdlib between 1.77.1 and 1.80.0, or the encoding_rs crate got bumped and changed in a way that exposes this issue whereas before it papered over it.