Support stripping a prefix from all file paths
This introduces the CLI option `--strip-path-prefix prefix` and strips the provided `prefix` from all paths as seen by content, in stack traces, and the debugger.

Previously, we unilaterally stripped the directory part of the top-level script's path instead. This new approach is better in two ways:
1. it works for files that are in ancestor or sibling directories to the top-level script
2. it enables using an ancestor directory as the root

The second of these advantages can be seen in the diffs to the test expectations files: the stacks for those are more informative, because they now include the directory of the actual test, since we're only stripping the `[StarlingMonkey root]/tests` part of the path.