Update OpenSSL to version 3.0.16 (#233)
This updates OpenSSL to the latest version of the 3.0 release train, from the previous, very old, 3.0.7.

It also moves to using CPM to download and patch the OpenSSL source code, which speeds up building different configurations from the same source tree, and eliminates some issues where reconfiguring caused issues with applying the patches.