Port SpiderMonkey build integration to CMake
The motivation is that I wanted the resulting library to really just be a single static archive file, instead of an archive plus a loose set of object files. That meant adding a post-processing step, which now needed invalidation logic. All much easier to do in CMake than the shell script we had before.