Run all test types in the same release or debug build in CI
This merges the jobs for testing E2E & Integration tests and the WPT suite into a single job. Doing separate builds for each of these is wasteful, since the WPT-compatible builds can fully be used to also run the normal tests.

(I plan on removing the special build requirement for WPT in the near future, so any concerns around not running tests in the runtime configuration we actually want people to use should be short-lived.)