Try fixing release-please permissions
Let's see if this fixes the release permissions. It at least matches the [release-please action's documentation](https://github.com/googleapis/release-please-action#basic-configuration).

In looking at this, I also noticed that the published artifacts aren't quite what I expected—my apologies for not catching this earlier! There's a key difference between the `starling-raw.wasm` and `starling.wasm` builds, and we shouldn't rename the former to the latter for the released artifacts. Instead, I simply added both, so that they can both be made use of directly.