Update Rust bindings integration
This is a large commit with a grab-bag of changes needed (or nice to have) to get this PR to work on top of #255.

The bindings are still pretty rough, but the result passes the test suite, so at least nothing existing is broken by it.

I'd like to get this landed soon-ish and iterate on things from here.

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>