feat(dx): add default message for missing handler
This commit adds a default message for the HTTP handler to the 500
that is returned when the event loop has run but the request is still
not being responded to via the fetch-event integration.

In these cases, the user would normally receive a 500 with no
information on what went wrong (e.g. if run with `wasmtime serve`), so
with this PR we can do a little more to point towards the issue.