Unify release-please CI workflows
The previous approach didn't work as I thought it would. I think/hope that this approach will create separate PRs for StarlingMonkey and debugger releases.

One thing to note is that this (as the previous version) changes the StarlingMonkey releases to have a `starlingmonkey_` prefix. That seems like the right thing to me, but it does mean that the release artifact paths are changing.

Signed-off-by: Till Schneidereit <till@tillschneidereit.net>